+++ src/components/inputs/SearchBar/SearchBar.tsx
@@ -1,4 +1,4 @@
-import searchIcon from '../../assets/images/DarkMode/SearchIcon.png';
+import searchIcon from '../../../assets/images/DarkMode/SearchIcon.png';
 
 type searchBar = {
   searchButtonAction: any;
