+++ src/components/layout/Navbar/Navbar.tsx
@@ -1,4 +1,4 @@
-import winston from '../assets/images/DarkMode/winston-logo.png';
+import winston from '../../../assets/images/DarkMode/winston-logo.png';
 import ConnectButton from '../../inputs/buttons/ConnectButton/ConnectButton';
 import { Link } from 'react-router-dom';
